TAJO-2073: Upgrade parquet-mr to 1.8.1.

+++ tajo-storage/tajo-storage-hdfs/src/main/java/org/apache/tajo/storage/text/DelimitedTextFile.java
@@ -268,7 +268,7 @@ public class DelimitedTextFile {
 
     private final long endOffset;
     /** The number of actual read records */
-    private int recordCount = 0;
+    private long recordCount = 0;
 
     private DelimitedLineReader reader;
     private TextLineDeserializer deserializer;
